src/test/java/com/leetcode/math/easy/CountPrimes_204Test.java
package com.leetcode.math.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class CountPrimes_204Test {
  private final CountPrimes_204 solution = new CountPrimes_204();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(10, 4), arguments(0, 0), arguments(1, 0));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void countPrimes(int num, int expectedResult) {
    int actualResult = solution.countPrimes(num);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
